Correcion de servicios y errores

+++ src/main/java/com/uptc/frw/jpa/repository/CesionRepository.java
@@ -3,7 +3,7 @@ package com.uptc.frw.jpa.repository;
 import com.uptc.frw.jpa.model.Cesion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CesionRepository extends JpaRepository<Cesion, Long> {
+public interface CesionRepository extends JpaRepository<Cesion, Integer> {
 }
