+++ src/main/java/com/uptc/frw/jpa/repository/CompraRepository.java
@@ -3,5 +3,5 @@ package com.uptc.frw.jpa.repository;
 import com.uptc.frw.jpa.model.Compra;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompraRepository extends JpaRepository<Compra, Long> {
+public interface CompraRepository extends JpaRepository<Compra, Integer> {
 }
